+++ R/plot.nutr_input.R
@@ -15,7 +15,7 @@
 #'
 #' @examples
 #' nutrients_input <- simulate_nutrient_sine(n = 3, max_i = 4380, input_mn = 1,
-#' frequency = 3, amplitude_sd = 0.5)
+#' frequency = 3, noise = 0.5)
 #' plot(nutrients_input)
 #'
 #' @aliases plot.nutr_input