+++ R/simulate_nutrient_noise.R
@@ -7,8 +7,7 @@
 #' @param max_i Integer with maximum number of simulation time steps.
 #' @param frequency Numeric frequency.
 #' @param input_mn Numeric with mean input amount and frequency.
-#' @param amplitude_mn Numeric with mean amplitude and phase.
-#' @param noise_sd Numeric with noise function sd.
+#' @param noise Numeric with noise function sd.
 #' @param n_noise Integer with number of noise functions
 #' @param verbose If TRUE, progress reports are printed.
 #'
@@ -18,8 +17,8 @@
 #' @return nutr_input
 #'
 #' @examples
-#' nutrients_input <- simulate_nutrient_noise(n = 3, max_i = 4380, frequency = 5, input_mn = 1,
-#' amplitude_mn = 0.5, noise_sd = 1.0)
+#' nutrients_input <- simulate_nutrient_noise(n = 3, max_i = 4380, frequency = 5,
+#' input_mn = 1, noise = 0.5)
 #' plot(nutrients_input)
 #'
 #' @aliases simulate_nutrient_noise
@@ -27,20 +26,12 @@
 #'
 #' @export
 simulate_nutrient_noise <- function(n, max_i, frequency = 0.0, input_mn = 0.0,
-                                    amplitude_mn = 0.0, noise_sd = 0.0, n_noise = 3,
-                                    verbose = TRUE) {
+                                    noise = 0.0, n_noise = 3, verbose = TRUE) {
 
-  # check amplitude argument
-  if (amplitude_mn < 0.0 || amplitude_mn > 1.0) {
+  # check noise argument
+  if (noise < 0.0 || noise > 1.0) {
 
-    stop("'amplitude_mn' should be 0.0 <= x <= 1.0", call. = FALSE)
-
-  }
-
-  # check noise_sd argument
-  if (noise_sd < 0.0 || noise_sd > 1.0) {
-
-    stop("'noise_sd' should be 0.0 <= x <= 1.0", call. = FALSE)
+    warning("'noise' should be 0.0 <= x <= 1.0", call. = FALSE)
 
   }
 
@@ -57,19 +48,22 @@ simulate_nutrient_noise <- function(n, max_i, frequency = 0.0, input_mn = 0.0,
 
     for (j in 1:n_noise) {
 
-      # draw random numbers from norm distribution
-      amplitude_mod <- arrR:::rcpp_rnorm(mean = amplitude_mn, sd = noise_sd, min = 0.0, max = 1.0)
-
-      period_mod <- arrR:::rcpp_rnorm(mean = 1.0, sd = noise_sd, min = 0.0, max = Inf)
-
-      # calculate period for number of input peaks (period = 2 * pi / b)
-      period_temp <- (frequency / (max_i / (2 * pi))) * period_mod
+      # draw random numbers from norm distribution to modify amplitude amplitude_mn
+      amplitude_mod <- stats::rnorm(n = 1, mean = 0.0, sd = noise) # *
+        # sample(x = c(-1, 1), size = 1, prob = c(noise * (0.5 / 1), 1 - noise * (0.5 / 1)))
 
       # calculate values for sine curve
       amplitude_temp <- input_mn * amplitude_mod
 
+      # draw random number from norm distribution to modify phase
+      period_mod <- stats::rnorm(n = 1, mean = 1.0, sd = noise)
+
+      # calculate period for number of input peaks (period = 2 * pi / b)
+      period_temp <- (frequency / (max_i / (2 * pi))) * period_mod
+
       # calculate phase
-      phase_temp <- 0.0 # ifelse(test = period_temp == 0, yes = 0.0, no = ((2 * pi) / period_temp) * phase_mod)
+      phase_temp <- 0.0
+      # ifelse(test = period_temp == 0, yes = 0.0, no = ((2 * pi) / period_temp) * phase_mod)
 
       # calculate sine curve; vertical shift to make sure x > 0
       # amplitude * sin(period * (x + phase)) + vertical
@@ -81,9 +75,11 @@ simulate_nutrient_noise <- function(n, max_i, frequency = 0.0, input_mn = 0.0,
     values_temp <- apply(X = do.call("cbind", input_temp), MARGIN = 1, FUN = "mean")
 
     # check if any values are below zero or above amplitude
-    if (any(values_temp < 0) && verbose) {
+    if (any(values_temp < 0) || any(values_temp > 2 * input_mn)) {
 
-      stop("Negative input value created. Please check arguments.", call. = FALSE)
+      values_temp <- scales::rescale(values_temp, to = c(0, 2 * input_mn))
+
+      if (verbose) warning("Rescaling nutrient inputs values to be 0 <= x <= 2 * input_mn.", call. = FALSE)
 
     }